Bug fix: Separate Default producer and consumer connections to allow recovery from flow control
In flow state, RMQ blocks the publishing connection. If consumers are using same connection,
they will be unable to drain the queue, thus defeating the whole purpose of flow control.
Keeping connections separate will allow only publishers to be blocked when rmq triggers flow state,
and for consumers to drain the queues and recover the queue eventually by adding credits.