Bugfix: removing check from create connection function
This check was supposed to validate that user is not setting custom connection
name in actorconfig same as one of default connections.

Therefore it needed to be done much earlier in the flow, ideally as a validation method
in the actor config class itself. At its current location, it is blocking creation of even default connections.
So, removing this check from here is the right thing to do.
We haven't added the check as a config validation of actor config since the default connection
names defined in Constants file, now start with "__" prefix and it is practically impossible
for a user to define same connection name in actor config.